feat: 예약 정보와 예약 요청 정보를 dto로 분리
