+++ src/main/java/roomescape/ReservationDto.java
@@ -4,31 +4,15 @@ import java.time.LocalDate;
 import java.time.LocalTime;
 
 public class ReservationDto {
-    private final long id;
     private final String name;
     private final LocalDate date;
     private final LocalTime time;
 
-    public ReservationDto(long id, String name, LocalDate date, LocalTime time) {
-        this.id = id;
+    public ReservationDto(String name, LocalDate date, LocalTime time) {
         this.name = name;
         this.date = date;
         this.time = time;
     }
 
-    public long getId() {
-        return id;
-    }
 
-    public String getName() {
-        return name;
-    }
-
-    public LocalDate getDate() {
-        return date;
-    }
-
-    public LocalTime getTime() {
-        return time;
-    }
 }
